[rx2] move test packages into same structure as dbflow packages so we can access package private methods. Add mockito kotlin. fix cursor result rx stream not working for max long.

+++ dbflow/src/main/java/com/raizlabs/android/dbflow/sql/language/CursorResult.java
@@ -136,7 +136,7 @@ public class CursorResult<TModel> implements IFlowCursorIterator<TModel> {
     }
 
     @Override
-    public FlowCursorIterator<TModel> iterator(int startingLocation, int limit) {
+    public FlowCursorIterator<TModel> iterator(int startingLocation, long limit) {
         return new FlowCursorIterator<>(this, startingLocation, limit);
     }
